requirements.txt
keyboard>=0.13.5
Pillow>=10.0.0
mss>=9.0.0
pywin32>=306
pystray>=0.19.0
